reading_pdf_parragraphs/requirements.txt
pymupdf==1.24.9
boto3==1.34.154
PyMuPDF==1.24.9
python-dotenv>=1.0.1

# httpx>=0.27.2
# tqdm>=4.66.4
# dspy-ai
# litellm[proxy]